requirements.txt
requests
pyyaml
scikit-learn
prometheus-api-client
gitpython
